refactor(web): extract useModelsAndLoras hook from App.jsx (sc-1651)
Fourth Phase-A slice. Models and LoRAs share one hook (a LoRA import/delete
re-pulls both via the lora overlay). hooks/useModelsAndLoras.js owns models + loras
state and refreshLoras + deleteModel/deleteLora + createModel/LoraImportJob +
createModelDownloadJob/createModelConvertJob (plus the upload-size constants that
only those functions used). App keeps the cross-cutting orchestrators — refreshData
(bulk loader; seeds models+loras through the hook's setters) and
refreshDataWithLoraOverlay (refreshData + refreshLoras, also called by the SSE
handler) — and passes them into the hook. imageModels/videoModels useMemos stay in
App, deriving from the hook's models. No screen prop-signature changes.

Behavior-preserving: App.jsx 2026 -> 1893 lines (god module 2334 -> 1893, -441 over
four slices). 107 web vitest pass, build clean, browser smoke (Rust API on a temp
data dir): the model catalog + LoRA families render (refreshData -> hook setters),
and a model Download created a queued job through the hook (optimistic setJobs ->
Queue 1 -> refreshData) with no console errors.
